Problem4and8
I know it's a LOT of code, but I made the graph creation algorithm before I thought too hard about how searching the graphs would be coded, so it's probably a little more complex than it needs to be. Luckily, our programs for this course are so small that runtime will likely not be noticeably long.

Also, I placed those last two lines just to allow user to examine all 10 graphs before continuing or exiting the program, depending on how our final product is set up. So those two lines will likely need to be removed/edited a little.